Refactor the api
